refactor(auth): 로그인·회원가입을 feature-sliced 구조로 재편
login-page/register-page 플랫 파일을 features/login·features/register 로 분리하고
각 feature 를 ui/lib/remote 단위로 쪼갠다. public API(index.ts)는 동일해
admin·shorts import 는 변경 없음.

- shared/: config/paths, lib/cn, lib/extract-message
- features/login: ui/login-page · lib/schema · remote/use-login
- features/register: ui(register-page·signup-form·signup-profile-form·
  additional-fields-section·auth-shell·field·text-input·radio-card·submit-button)
  · lib(additional·schema) · remote(use-signup·use-signup-profile)
- remote 훅은 hook 시그니처를 미러링해 axios(withCredentials/Authorization)
  기본값만 주입하고 제네릭(에러 타입 등)은 호출부에서 유지되게 함.
