+++ packages/auth/src/features/login/index.ts
@@ -1,0 +1,1 @@
+export { LoginPage } from "./ui/login-page";
