+++ packages/auth/src/features/register/index.ts
@@ -1,0 +1,1 @@
+export { Register } from "./ui/register-page";
