+++ packages/auth/src/features/register/ui/auth-shell.tsx
@@ -1,0 +1,21 @@
+export const AuthShell = ({
+  title,
+  description,
+  children,
+}: {
+  title: string;
+  description: string;
+  children: React.ReactNode;
+}) => (
+  <div className="min-h-screen bg-white text-slate-900">
+    <div className="mx-auto flex min-h-screen max-w-lg items-center px-6 py-12">
+      <div className="w-full rounded-2xl border border-slate-200 bg-white p-8 shadow-lg shadow-slate-200/60">
+        <div className="space-y-2">
+          <h1 className="text-2xl font-semibold tracking-tight">{title}</h1>
+          <p className="text-sm text-slate-500">{description}</p>
+        </div>
+        {children}
+      </div>
+    </div>
+  </div>
+);
