+++ packages/auth/src/shared/config/paths.ts
